order-consumer/requirements.txt
pika==1.3.1
requests==2.31.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-instrumentation-requests>=0.41b0
